Change signer methods to SigningUtility
