+++ src/NuGet.Core/NuGet.Commands/SignCommand/SignCommandRunner.cs
@@ -97,10 +97,8 @@ namespace NuGet.Commands
 
                         // Set the output of the signing operation to a temp package because signing is cannot be done in place
                         var tempPackagePath = Path.GetTempFileName();
-                        using (var signerOptions = new SignerOptions(packagePath, tempPackagePath, overwrite, signatureProvider, signPackageRequest, logger))
-                        {
-                            await SignPackageAsync(signerOptions, outputPath, token);
-                        }
+                        var signingOptions = new SigningOptions(packagePath, tempPackagePath, overwrite, signatureProvider, logger);
+                        await SignPackageAsync(signingOptions, signPackageRequest, outputPath, token);
                     }
                     catch (Exception e)
                     {
@@ -131,21 +129,20 @@ namespace NuGet.Commands
         }
 
         private async Task<int> SignPackageAsync(
-            SignerOptions signerOptions,
+            SigningOptions signingOptions,
+            SignPackageRequest signRequest,
             string packageOutputPath,
             CancellationToken token)
         {
             try
             {
-                var signer = new Signer(signerOptions);
-
-                await signer.SignAsync(token);
+                await SigningUtility.SignAsync(signingOptions, signRequest, token);
             }
             finally
             {
-                if (File.Exists(signerOptions.OutputFilePath))
+                if (File.Exists(signingOptions.OutputFilePath))
                 {
-                    FileUtility.Replace(signerOptions.OutputFilePath, packageOutputPath);
+                    FileUtility.Replace(signingOptions.OutputFilePath, packageOutputPath);
                 }
             }
