+++ src/NuGet.Core/NuGet.Packaging/Signing/Authoring/SigningOptions.cs
@@ -8,10 +8,8 @@ using NuGet.Common;
 
 namespace NuGet.Packaging.Signing
 {
-    public sealed class SignerOptions : IDisposable
+    public sealed class SigningOptions
     {
-        private bool _isDisposed;
-
         /// <summary>
         /// Path to the package file that will be used as an input for any signer operation.
         /// <remarks>Signer operations cannot be done in place, therefore this value should be different than OutputFilePath.</remarks>
@@ -35,16 +33,11 @@ namespace NuGet.Packaging.Signing
         public ISignatureProvider SignatureProvider { get; }
 
         /// <summary>
-        /// Request to sign packages
-        /// </summary>
-        public SignPackageRequest SignRequest { get; }
-
-        /// <summary>
         /// Logger to be used to display the logs during the execution of signer actions.
         /// </summary>
         public ILogger Logger { get; }
 
-        public SignerOptions(string packageFilePath, string outputFilePath, bool overwrite, ISignatureProvider signatureProvider, SignPackageRequest signRequest, ILogger logger)
+        public SigningOptions(string packageFilePath, string outputFilePath, bool overwrite, ISignatureProvider signatureProvider, ILogger logger)
         {
             PackageFilePath = packageFilePath ?? throw new ArgumentNullException(nameof(packageFilePath));
             OutputFilePath = outputFilePath ?? throw new ArgumentNullException(nameof(outputFilePath));
@@ -55,7 +48,6 @@ namespace NuGet.Packaging.Signing
             }
 
             SignatureProvider = signatureProvider ?? throw new ArgumentNullException(nameof(signatureProvider));
-            SignRequest = signRequest ?? throw new ArgumentNullException(nameof(signRequest));
             Logger = logger ?? throw new ArgumentNullException(nameof(logger));
             Overwrite = overwrite;
         }
@@ -63,21 +55,6 @@ namespace NuGet.Packaging.Signing
         private static string NormalizeFilePath(string filePath)
         {
             return Path.GetFullPath(filePath).TrimEnd('\\');
-        }
-
-        /// <summary>
-        /// Disposes of this instance.
-        /// </summary>
-        public void Dispose()
-        {
-            if (!_isDisposed)
-            {
-                SignRequest?.Dispose();
-
-                GC.SuppressFinalize(this);
-
-                _isDisposed = true;
-            }
         }
     }
 }
