+++ src/NuGet.Core/NuGet.Packaging/Signing/Utility/SigningUtility.cs
@@ -3,11 +3,14 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Security.Cryptography;
 #if IS_DESKTOP
 using System.Security.Cryptography.Pkcs;
 #endif
 using System.Security.Cryptography.X509Certificates;
+using System.Threading;
+using System.Threading.Tasks;
 using NuGet.Common;
 
 namespace NuGet.Packaging.Signing
@@ -168,6 +171,88 @@ namespace NuGet.Packaging.Signing
 
             return signer;
         }
+
+        public static async Task RemoveSignatureAsync(string packageFilePath, string outputFilePath, CancellationToken token)
+        {
+            using (var packageReadStream = File.OpenRead(packageFilePath))
+            using (var packageWriteStream = File.Open(outputFilePath, FileMode.OpenOrCreate))
+            using (var package = new SignedPackageArchive(packageReadStream, packageWriteStream))
+            {
+                await package.RemoveSignatureAsync(token);
+            }
+        }
+
+        /// <summary>
+        /// Add a signature to a package.
+        /// </summary>
+        public static async Task SignAsync(SigningOptions options, SignPackageRequest signRequest, CancellationToken token)
+        {
+            token.ThrowIfCancellationRequested();
+
+            Verify(signRequest, options.Logger);
+
+            var inputPackagePath = options.PackageFilePath;
+            var tempPackagePath = Path.GetTempFileName();
+
+            try
+            {
+                using (var packageReadStream = File.OpenRead(inputPackagePath))
+                using (var packageWriteStream = File.Open(tempPackagePath, FileMode.OpenOrCreate))
+                using (var package = new SignedPackageArchive(packageReadStream, packageWriteStream))
+                {
+                    var primarySignature = await package.GetPrimarySignatureAsync(token);
+
+                    if (options.Overwrite && primarySignature != null)
+                    {
+                        await package.RemoveSignatureAsync(token);
+                        inputPackagePath = tempPackagePath;
+                    }
+                }
+
+                using (var packageReadStream = File.OpenRead(inputPackagePath))
+                using (var packageWriteStream = File.Open(options.OutputFilePath, FileMode.OpenOrCreate))
+                using (var package = new SignedPackageArchive(packageReadStream, packageWriteStream))
+                {
+                    if (await package.IsZip64Async(token))
+                    {
+                        throw new SignatureException(NuGetLogCode.NU3006, Strings.ErrorZip64NotSupported);
+                    }
+
+                    var hashAlgorithm = signRequest.SignatureHashAlgorithm;
+
+                    var zipArchiveHash = await package.GetArchiveHashAsync(hashAlgorithm, token);
+                    var signatureContent = GenerateSignatureContent(hashAlgorithm, zipArchiveHash);
+                    var signature = await options.SignatureProvider.CreatePrimarySignatureAsync(signRequest, signatureContent, options.Logger, token);
+
+                    using (var stream = new MemoryStream(signature.GetBytes()))
+                    {
+                        await package.AddSignatureAsync(stream, token);
+                    }
+                }
+            }
+            finally
+            {
+                FileUtility.Delete(tempPackagePath);
+            }
+        }
+
+        private static SignatureContent GenerateSignatureContent(Common.HashAlgorithmName hashAlgorithmName, byte[] zipArchiveHash)
+        {
+            var base64ZipArchiveHash = Convert.ToBase64String(zipArchiveHash);
+
+            return new SignatureContent(SigningSpecifications.V1, hashAlgorithmName, base64ZipArchiveHash);
+        }
+#else
+        /// <summary>
+        /// Remove a signature from a package.
+        /// </summary>
+        public static Task RemoveSignatureAsync(string packageFilePath, string outputFilePath, CancellationToken token) => throw new NotImplementedException();
+
+
+        /// <summary>
+        /// Add a signature to a package.
+        /// </summary>
+        public static Task SignAsync(SigningOptions options, SignPackageRequest signRequest, CancellationToken token) => throw new NotImplementedException();
 #endif
     }
 }
